Fully functioning minesweeper
check win has been updated and is now working, 
Issue:
Reveal cell didn't actually update the revealed cell board. It would only print the correct element

How I solved it:
I made a new setter method that would take the inputs of the row and col I would like to update within the revealed board. Then it would set the exact element that iIwanted to true so that right after it can print out on the board the element. This allowed for check win to work since the cell was updated on the reveal board.